+++ lib/lighthouseProvider/deviceExtensions/DeviceWithExtensions.dart
@@ -1,0 +1,6 @@
+import './DeviceExtensions.dart';
+
+/// An interface for devices that have extra function extensions.
+abstract class DeviceWithExtensions {
+  final Set<DeviceExtensions> deviceExtensions = Set();
+}