+++ lib/lighthouseProvider/devices/LighthouseV2Device.dart
@@ -5,6 +5,9 @@ import 'package:flutter_blue/flutter_blue.dart';
 import 'package:lighthouse_pm/lighthouseProvider/LighthousePowerState.dart';
 import 'package:lighthouse_pm/lighthouseProvider/ble/DefaultCharacteristics.dart';
 import 'package:lighthouse_pm/lighthouseProvider/ble/Guid.dart';
+import 'package:lighthouse_pm/lighthouseProvider/deviceExtensions/DeviceExtensions.dart';
+import 'package:lighthouse_pm/lighthouseProvider/deviceExtensions/DeviceWithExtensions.dart';
+import 'package:lighthouse_pm/lighthouseProvider/deviceExtensions/IdentifyDeviceExtension.dart';
 import 'package:lighthouse_pm/lighthouseProvider/devices/BLEDevice.dart';
 import 'package:lighthouse_pm/lighthouseProvider/helpers/FlutterBlueExtensions.dart';
 
@@ -16,10 +19,17 @@ const String _POWER_CHARACTERISTIC = '00001525-1212-efde-1523-785feabcd124';
 
 const String _CHANNEL_CHARACTERISTIC = '00001524-1212-EFDE-1523-785FEABCD124';
 
-class LighthouseV2Device extends BLEDevice {
-  LighthouseV2Device(BluetoothDevice device) : super(device);
+const String _IDENTIFY_CHARACTERISTIC = '00008421-1212-EFDE-1523-785FEABCD124';
 
+class LighthouseV2Device extends BLEDevice implements DeviceWithExtensions {
+  LighthouseV2Device(BluetoothDevice device) : super(device) {
+    deviceExtensions.add(IdentifyDeviceExtension(onTap: identify));
+  }
+
+  @override
+  final Set<DeviceExtensions> deviceExtensions = Set();
   BluetoothCharacteristic /* ? */ _characteristic;
+  BluetoothCharacteristic /* ? */ _identifyCharacteristic;
   String /* ? */ _firmwareVersion;
   final Map<String, String> _otherMetadata = Map();
   static const _SUPPORTED_CHARACTERISTICS_LIST = const [
@@ -94,6 +104,8 @@ class LighthouseV2Device extends BLEDevice {
         LighthouseGuid.fromString(_POWER_CHARACTERISTIC);
     final channelCharacteristic =
         LighthouseGuid.fromString(_CHANNEL_CHARACTERISTIC);
+    final identifyCharacteristic =
+        LighthouseGuid.fromString(_IDENTIFY_CHARACTERISTIC);
 
     debugPrint('Finding service for device: ${this.deviceIdentifier}');
     final List<BluetoothService> services =
@@ -105,6 +117,11 @@ class LighthouseV2Device extends BLEDevice {
 
         if (uuid == powerCharacteristic) {
           this._characteristic = characteristic;
+          continue;
+        }
+        if (uuid == identifyCharacteristic) {
+          this._identifyCharacteristic = characteristic;
+          continue;
         }
         if (uuid == channelCharacteristic) {
           try {
@@ -114,6 +131,7 @@ class LighthouseV2Device extends BLEDevice {
             debugPrint('Unable to get channel because: $e');
             debugPrint('$s');
           }
+          continue;
         }
 
         if (DefaultCharacteristics.FIRMWARE_REVISION_CHARACTERISTIC
@@ -128,6 +146,7 @@ class LighthouseV2Device extends BLEDevice {
             debugPrint('Unable to get firmware version because: $e');
             debugPrint('$s');
           }
+          continue;
         }
         for (final defaultCharacteristic in _SUPPORTED_CHARACTERISTICS_LIST) {
           if (defaultCharacteristic.isEqualToGuid(uuid)) {
@@ -184,5 +203,17 @@ class LighthouseV2Device extends BLEDevice {
           break;
       }
     }
+  }
+
+  /// Identify the current lighthouse by way of blinking the front LED.
+  Future<void> identify() async {
+    if (this._identifyCharacteristic == null) {
+      debugPrint('No identify characteristic set!');
+      return;
+    }
+    await transactionMutex.acquire();
+    // Write any byte to the characteristic to start the identify option.
+    await this._identifyCharacteristic.write([0x00]);
+    transactionMutex.release();
   }
 }